feat: Show how to implement outside of client
