+++ packages/core/src/client.ts
@@ -63,13 +63,10 @@ import { parseSampleRate } from './utils/parseSampleRate';
 import { prepareEvent } from './utils/prepareEvent';
 import { showSpanDropWarning, spanToTraceContext } from './utils/spanUtils';
 import { convertSpanJsonToTransactionEvent, convertTransactionEventToSpanJson } from './utils/transactionEvent';
-import type { Log, SerializedOtelLog } from './types-hoist/log';
-import { SEVERITY_TEXT_TO_SEVERITY_NUMBER, createOtelLogEnvelope, logAttributeToSerializedLogAttribute } from './log';
 import { _getSpanForScope } from './utils/spanOnScope';
 
 const ALREADY_SEEN_ERROR = "Not capturing exception because it's already been captured.";
 const MISSING_RELEASE_FOR_SESSION_ERROR = 'Discarded session because of missing or non-string release';
-const MAX_LOG_BUFFER_SIZE = 100;
 
 /**
  * Base implementation for all JavaScript SDK clients.
@@ -125,8 +122,6 @@ export abstract class Client<O extends ClientOptions = ClientOptions> {
   // eslint-disable-next-line @typescript-eslint/ban-types
   private _hooks: Record<string, Function[]>;
 
-  private _logsBuffer: Array<SerializedOtelLog>;
-
   /**
    * Initializes this client instance.
    *
@@ -139,7 +134,6 @@ export abstract class Client<O extends ClientOptions = ClientOptions> {
     this._outcomes = {};
     this._hooks = {};
     this._eventProcessors = [];
-    this._logsBuffer = [];
 
     if (options.dsn) {
       this._dsn = makeDsn(options.dsn);
@@ -268,58 +262,6 @@ export abstract class Client<O extends ClientOptions = ClientOptions> {
   public captureCheckIn?(checkIn: CheckIn, monitorConfig?: MonitorConfig, scope?: Scope): string;
 
   /**
-   * Captures a log event and sends it to Sentry.
-   *
-   * @param log The log event to capture.
-   *
-   * @experimental This method will experience breaking changes. This is not yet part of
-   * the stable Sentry SDK API and can be changed or removed without warning.
-   */
-  public captureLog({ level, message, attributes, severityNumber }: Log, currentScope = getCurrentScope()): void {
-    const { _experiments, release, environment } = this.getOptions();
-    if (!_experiments?.enableLogs) {
-      DEBUG_BUILD && logger.warn('logging option not enabled, log will not be captured.');
-      return;
-    }
-
-    const [, traceContext] = _getTraceInfoFromScope(this, currentScope);
-
-    const logAttributes = {
-      ...attributes,
-    };
-
-    if (release) {
-      logAttributes.release = release;
-    }
-
-    if (environment) {
-      logAttributes.environment = environment;
-    }
-
-    const span = _getSpanForScope(currentScope);
-    if (span) {
-      // Add the parent span ID to the log attributes for trace context
-      logAttributes['sentry.trace.parent_span_id'] = span.spanContext().spanId;
-    }
-
-    const serializedLog: SerializedOtelLog = {
-      severityText: level,
-      body: {
-        stringValue: message,
-      },
-      attributes: Object.entries(logAttributes).map(([key, value]) => logAttributeToSerializedLogAttribute(key, value)),
-      timeUnixNano: `${new Date().getTime().toString()}000000`,
-      traceId: traceContext?.trace_id,
-      severityNumber: severityNumber ?? SEVERITY_TEXT_TO_SEVERITY_NUMBER[level],
-    };
-
-    this._logsBuffer.push(serializedLog);
-    if (this._logsBuffer.length > MAX_LOG_BUFFER_SIZE) {
-      this._flushLogsBuffer();
-    }
-  }
-
-  /**
    * Get the current Dsn.
    */
   public getDsn(): DsnComponents | undefined {
@@ -358,7 +300,6 @@ export abstract class Client<O extends ClientOptions = ClientOptions> {
    * still events in the queue when the timeout is reached.
    */
   public flush(timeout?: number): PromiseLike<boolean> {
-    this._flushLogsBuffer();
     const transport = this._transport;
     if (transport) {
       this.emit('flush');
@@ -1195,21 +1136,6 @@ export abstract class Client<O extends ClientOptions = ClientOptions> {
 
     const envelope = createClientReportEnvelope(outcomes, this._options.tunnel && dsnToString(this._dsn));
 
-    // sendEnvelope should not throw
-    // eslint-disable-next-line @typescript-eslint/no-floating-promises
-    this.sendEnvelope(envelope);
-  }
-
-  /**
-   * Flushes the logs buffer to Sentry.
-   */
-  protected _flushLogsBuffer(): void {
-    if (this._logsBuffer.length === 0) {
-      return;
-    }
-
-    const envelope = createOtelLogEnvelope(this._logsBuffer, this._options._metadata, this._options.tunnel, this._dsn);
-    this._logsBuffer = [];
     // sendEnvelope should not throw
     // eslint-disable-next-line @typescript-eslint/no-floating-promises
     this.sendEnvelope(envelope);
